30 mins slot booking allowed in conference

+++ src/main/java/com/letswork/crm/entities/BookConferenceRoom.java
@@ -36,7 +36,7 @@ public class BookConferenceRoom extends Base{
 	
 	private Integer numberOfGuests;
 		
-	private Integer numberOfHours;
+	private Float numberOfHours;
 	
 	private String email;
